Move storage implementation to internal
This may need to come out of internal if we ever need to support a third party storage that lives outside of this repo, but at the moment this can safely be moved to internal to delay any API concerns. Towards #280.
